Implement heartbeat and show online guild members

+++ DiscordVoice/Models/HelloPayload.swift
@@ -7,8 +7,9 @@
 
 import Foundation
 
+/// https://discord.com/developers/docs/topics/gateway#hello
 struct HelloPayload: Codable {
-    let heartbeatInterval: Int
+    let heartbeatInterval: Double // milliseconds
     
     enum CodingKeys: String, CodingKey {
         case heartbeatInterval = "heartbeat_interval"
